Removed docker compose volume
The volume granted live reloading, but I believe it's causing a bug
where launching the containers with docker-compose up --build fails. The
command fails until npm install is ran in the host environment. The
volume seems to overwrite the build of the Dockerfile with the contents
of the host machine. Instead, the build from the Dockerfile that
includes npm install and npm start commands should successfully run.